requirements.txt
PyYAML
plotly
plotly.express
matplotlib
nbformat
dash
openpyxl
requests
python-dateutil
datetime
streamlit
